feat: long 타입으로 수정

+++ src/main/java/kr/hhplus/be/server/domain/point/service/PointService.java
@@ -20,7 +20,7 @@ public class PointService {
     private final PointRepository pointRepository;
 
     //포인트 조회
-    public PointResponse getUserPoint(int userId){
+    public PointResponse getUserPoint(long userId){
         Point point = pointRepository.getUserPoint(userId);
         return new PointResponse(userId, point.getPoint());
     }
